# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export interface SendAnthropicMessageBody {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export interface ForexAnalysisRequest {
  pair: string;
  timeframe: string;
  price: number;
  bias?: string;
  context?: string;
}

export interface ConfluenceRequest {
  pair: string;
  timeframe: string;
  price: number;
  trend?: string;
  newsContext?: string;
}

export type ConfluenceResponseFactorsItem = {
  label: string;
  value: number;
  description: string;
};

export type ConfluenceResponseTradeIdea = {
  pair: string;
  direction: string;
  entry: string;
  stopLoss: string;
  takeProfit1: string;
  takeProfit2: string;
  riskReward: string;
  probability: number;
  explanation: string;
};

export interface ConfluenceResponse {
  score: number;
  bias: string;
  factors: ConfluenceResponseFactorsItem[];
  tradeIdea: ConfluenceResponseTradeIdea;
  summary: string;
}

export type NewsResponseKeyEventsItem = {
  event: string;
  impact: string;
  direction: string;
};

export interface NewsResponse {
  pair: string;
  summary: string;
  sentiment: string;
  keyEvents: NewsResponseKeyEventsItem[];
  generatedAt: string;
}

export interface Trade {
  id: number;
  pair: string;
  direction: string;
  entryPrice: string;
  stopLoss: string;
  takeProfit: string;
  lotSize?: number | null;
  result?: string | null;
  pips?: number | null;
  riskReward?: string | null;
  strategy?: string | null;
  notes?: string | null;
  status: string;
  createdAt: string;
}

export interface CreateTradeBody {
  pair: string;
  direction: string;
  entryPrice: string;
  stopLoss: string;
  takeProfit: string;
  strategy?: string;
  notes?: string;
  lotSize?: number;
  status: string;
}

export interface UpdateTradeBody {
  result?: string;
  pips?: number;
  riskReward?: string;
  status?: string;
  notes?: string;
  pair?: string;
  direction?: string;
  entryPrice?: string;
  stopLoss?: string;
  takeProfit?: string;
  lotSize?: number;
}

export interface JournalStats {
  totalTrades: number;
  winRate: number;
  averageRR: number;
  netPips: number;
  winCount: number;
  lossCount: number;
  breakEvenCount: number;
}