# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /journal/trades:
    get:
      operationId: listTrades
      tags: [journal]
      summary: List all journal trades
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: createTrade
      tags: [journal]
      summary: Create a new trade entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTradeBody"
      responses:
        "201":
          description: Created trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /journal/trades/{id}:
    put:
      operationId: updateTrade
      tags: [journal]
      summary: Update a trade entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTradeBody"
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
    delete:
      operationId: deleteTrade
      tags: [journal]
      summary: Delete a trade entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: integer
        pair:
          type: string
        direction:
          type: string
        entryPrice:
          type: string
        stopLoss:
          type: string
        takeProfit:
          type: string
        actualExit:
          type: string
          nullable: true
        lotSize:
          type: number
          nullable: true
        result:
          type: string
          nullable: true
        pips:
          type: number
          nullable: true
        riskReward:
          type: string
          nullable: true
        strategy:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - pair
        - direction
        - entryPrice
        - stopLoss
        - takeProfit
        - status
        - createdAt
    CreateTradeBody:
      type: object
      properties:
        pair:
          type: string
        direction:
          type: string
        entryPrice:
          type: string
        stopLoss:
          type: string
        takeProfit:
          type: string
        actualExit:
          type: string
        lotSize:
          type: number
        strategy:
          type: string
        notes:
          type: string
        status:
          type: string
      required:
        - pair
        - direction
        - entryPrice
        - stopLoss
        - takeProfit
        - status
    UpdateTradeBody:
      type: object
      properties:
        result:
          type: string
        pips:
          type: number
        riskReward:
          type: string
        status:
          type: string
        notes:
          type: string
        pair:
          type: string
        direction:
          type: string
        entryPrice:
          type: string
        stopLoss:
          type: string
        takeProfit:
          type: string
        actualExit:
          type: string
...