# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListAnthropicConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListAnthropicConversationsResponse = zod.array(
  ListAnthropicConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateAnthropicConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetAnthropicConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAnthropicConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteAnthropicConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListAnthropicMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListAnthropicMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListAnthropicMessagesResponse = zod.array(
  ListAnthropicMessagesResponseItem,
);

/**
 * @summary Send a message and receive an AI response (SSE stream)
 */
export const SendAnthropicMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendAnthropicMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Get real AI-powered forex analysis for a pair
 */
export const AnalyzeForexBody = zod.object({
  pair: zod.string(),
  timeframe: zod.string(),
  price: zod.number(),
  bias: zod.string().optional(),
  context: zod.string().optional(),
});

/**
 * @summary Get confluence score and trade idea for a pair
 */
export const AnalyzeConfluenceBody = zod.object({
  pair: zod.string(),
  timeframe: zod.string(),
  price: zod.number(),
  trend: zod.string().optional(),
  newsContext: zod.string().optional(),
});

export const AnalyzeConfluenceResponse = zod.object({
  score: zod.number(),
  bias: zod.string(),
  factors: zod.array(
    zod.object({
      label: zod.string(),
      value: zod.number(),
      description: zod.string(),
    }),
  ),
  tradeIdea: zod.object({
    pair: zod.string(),
    direction: zod.string(),
    entry: zod.string(),
    stopLoss: zod.string(),
    takeProfit1: zod.string(),
    takeProfit2: zod.string(),
    riskReward: zod.string(),
    probability: zod.number(),
    explanation: zod.string(),
  }),
  summary: zod.string(),
});

/**
 * @summary Get AI-summarized fundamental news for a pair
 */
export const GetForexNewsParams = zod.object({
  pair: zod.coerce.string(),
});

export const GetForexNewsResponse = zod.object({
  pair: zod.string(),
  summary: zod.string(),
  sentiment: zod.string(),
  keyEvents: zod.array(
    zod.object({
      event: zod.string(),
      impact: zod.string(),
      direction: zod.string(),
    }),
  ),
  generatedAt: zod.string(),
});

/**
 * @summary List all journal trades
 */
export const ListTradesResponseItem = zod.object({
  id: zod.number(),
  pair: zod.string(),
  direction: zod.string(),
  entryPrice: zod.string(),
  stopLoss: zod.string(),
  takeProfit: zod.string(),
  actualExit: zod.string().nullish(),
  lotSize: zod.number().nullish(),
  result: zod.string().nullish(),
  pips: zod.number().nullish(),
  riskReward: zod.string().nullish(),
  strategy: zod.string().nullish(),
  notes: zod.string().nullish(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListTradesResponse = zod.array(ListTradesResponseItem);

/**
 * @summary Create a new trade entry
 */
export const CreateTradeBody = zod.object({
  pair: zod.string(),
  direction: zod.string(),
  entryPrice: zod.string(),
  stopLoss: zod.string(),
  takeProfit: zod.string(),
  actualExit: zod.string().optional(),
  strategy: zod.string().optional(),
  notes: zod.string().optional(),
  lotSize: zod.number().optional(),
  status: zod.string(),
});

/**
 * @summary Update a trade entry
 */
export const UpdateTradeParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTradeBody = zod.object({
  result: zod.string().optional(),
  pips: zod.number().optional(),
  riskReward: zod.string().optional(),
  status: zod.string().optional(),
  notes: zod.string().optional(),
  pair: zod.string().optional(),
  direction: zod.string().optional(),
  entryPrice: zod.string().optional(),
  stopLoss: zod.string().optional(),
  takeProfit: zod.string().optional(),
  actualExit: zod.string().optional(),
  lotSize: zod.number().optional(),
});

export const UpdateTradeResponse = zod.object({
  id: zod.number(),
  pair: zod.string(),
  direction: zod.string(),
  entryPrice: zod.string(),
  stopLoss: zod.string(),
  takeProfit: zod.string(),
  actualExit: zod.string().nullish(),
  lotSize: zod.number().nullish(),
  result: zod.string().nullish(),
  pips: zod.number().nullish(),
  riskReward: zod.string().nullish(),
  strategy: zod.string().nullish(),
  notes: zod.string().nullish(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a trade entry
 */
export const DeleteTradeParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get trade journal statistics
 */
export const GetJournalStatsResponse = zod.object({
  totalTrades: zod.number(),
  winRate: zod.number(),
  averageRR: zod.number(),
  netPips: zod.number(),
  winCount: zod.number(),
  lossCount: zod.number(),
  breakEvenCount: zod.number(),
});